Reorganize internal/cli files along dispatcher/runner/core seams
Prepare the physical package split (clicore/dispatcher/projectrunner) by
regrouping files inside package cli with no behavior change:

- Consolidate all command name constants into command_registry.go so the
  dispatcher-side registry no longer reaches into runner-only files
- Split tools.go into tool_catalog.go, tool_options.go, tool_params.go,
  global_args.go, runner_version.go, and dispatcher_version.go
- Split run_dispatch.go into dispatcher_process.go and runner_commands.go
- Move version writers next to their identity vars, argument predicates
  into global_args.go, tool settings loading into tool_settings.go, and
  internalCLIError into error_envelope.go
