+++ cli/internal/cli/compile_wait.go
@@ -12,7 +12,6 @@ import (
 )
 
 const (
-	compileCommandName        = "compile"
 	compileStatusCommandName  = "get-compile-status"
 	compileRequestIDParam     = "RequestId"
 	compileWaitParam          = domainReloadWaitParam
