+++ cli/internal/cli/error_envelope.go
@@ -302,3 +302,18 @@ func isSafeRetryCommand(command string) bool {
 		return false
 	}
 }
+
+func internalCLIError(message string, context errorContext) cliError {
+	return cliError{
+		ErrorCode:   errorCodeInternalError,
+		Phase:       errorPhaseExecution,
+		Message:     message,
+		Retryable:   false,
+		SafeToRetry: false,
+		ProjectRoot: context.projectRoot,
+		Command:     context.command,
+		NextActions: []string{
+			"Read the message and fix the local environment or command input before retrying.",
+		},
+	}
+}
