+++ cli/internal/cli/run_help.go
@@ -13,27 +13,6 @@ const (
 	maxCommandListDescriptionLength = 96
 )
 
-func isVersionRequest(args []string) bool {
-	return len(args) == 1 && (args[0] == "--version" || args[0] == "-v")
-}
-
-func isVersionJSONRequest(args []string) bool {
-	return len(args) == 2 && (args[0] == "--version" || args[0] == "-v") && args[1] == "--json"
-}
-
-func isHelpRequest(args []string) bool {
-	return len(args) == 1 && (args[0] == "--help" || args[0] == "-h")
-}
-
-func containsHelpRequest(args []string) bool {
-	for _, arg := range args {
-		if arg == "--help" || arg == "-h" {
-			return true
-		}
-	}
-	return false
-}
-
 func printHelp(stdout io.Writer) {
 	printMainHelp(
 		stdout,
