+++ cli/internal/cli/skills_targets.go
@@ -1,7 +1,6 @@
 package cli
 
 import (
-	"encoding/json"
 	"os"
 	"path/filepath"
 	"strings"
@@ -132,23 +131,6 @@ func removeEmptyDir(path string) error {
 	return os.Remove(path)
 }
 
-func loadDisabledTools(projectRoot string) []string {
-	settingsPath := filepath.Join(projectRoot, uloopSettingsDir, toolSettingsFile)
-	content, err := os.ReadFile(settingsPath)
-	if err != nil || len(strings.TrimSpace(string(content))) == 0 {
-		return []string{}
-	}
-
-	settings := toolSettingsData{}
-	if err := json.Unmarshal(content, &settings); err != nil {
-		return []string{}
-	}
-	if settings.DisabledTools == nil {
-		return []string{}
-	}
-	return settings.DisabledTools
-}
-
 func isSkillDisabledByToolSettings(skill skillDefinition, disabledTools []string) bool {
 	if len(disabledTools) == 0 {
 		return false
@@ -161,16 +143,4 @@ func isSkillDisabledByToolSettings(skill skillDefinition, disabledTools []string
 		return false
 	}
 	return isToolDisabledByToolSettings(toolName, disabledTools)
-}
-
-func isToolDisabledByToolSettings(toolName string, disabledTools []string) bool {
-	if len(disabledTools) == 0 {
-		return false
-	}
-	for _, disabledTool := range disabledTools {
-		if disabledTool == toolName {
-			return true
-		}
-	}
-	return false
 }
